# web/openapi.yaml
basePath: /api/v1
definitions:
...
    profile.ProfileCategoryPaginatedResponse:
        properties:
            categories:
                items:
                    $ref: '#/definitions/profile.ProfileCategoryResponse'
                type: array
            pagination:
                additionalProperties: true
                type: object
        type: object
    profile.ProfileCategoryResponse:
        properties:
            created_at:
                type: string
            id:
                type: string
            name:
                type: string
            sort_order:
                type: integer
            updated_at:
                type: string
            usage_count:
                type: integer
        type: object
...
    response.ErrorResponse:
        properties:
            error:
                type: string
            success:
                type: boolean
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: REST API for the Desa (village) information system. Implements Clean Architecture with strict dependency rules, JWT (HS256) authentication, Casbin RBAC, PostgreSQL persistence, and rate limiting.
    title: Desa API
    version: "1.0"
paths:
...
    /profile/categories:
        post:
            consumes:
                - application/json
            description: 'RBAC: profile:write.'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/profile.ProfileCategoryResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Create profile category (admin)
            tags:
                - profile-categories
    /profile/categories/{id}:
        delete:
            description: 'Returns 409 if category is in use. RBAC: profile:write.'
            parameters:
                - description: Category UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "204":
                    description: No Content
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "404":
                    description: Not Found
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "409":
                    description: Conflict
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Delete profile category (admin)
            tags:
                - profile-categories
...
    /public/profile/categories:
        get:
            description: With q search for autocomplete.
            parameters:
                - description: Search
                  in: query
                  name: q
                  type: string
                - description: Page
                  in: query
                  name: page
                  type: integer
                - description: Page size
                  in: query
                  name: limit
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/profile.ProfileCategoryPaginatedResponse'
                "400":
                    description: Bad Request
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
                "429":
                    description: Too Many Requests
                    schema:
                        $ref: '#/definitions/response.ErrorResponse'
            summary: List profile categories (public)
            tags:
                - profile-categories
...
schemes:
    - http
    - https
securityDefinitions:
    BearerAuth:
        description: 'JWT bearer token. Format: `Bearer {token}`'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...